Remove comment about speed from TotW 61 (Default Member Initializers)
It feels like a bummer, to start the Conclusion by saying that such initializers "won't make your program any faster". If that would really be an important conclusion of TotW 61, the main text of the item should elaborate on this aspect. Otherwise, it should not be part of the take-home message.